requirements/local.requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/local.requirements.txt dummy.requirements.in
#
anyio==3.7.0
    # via jupyter-server
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
argon2-cffi==21.3.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.2.3
    # via isoduration
asttokens==2.2.1
    # via stack-data
async-lru==2.0.2
    # via jupyterlab
attrs==23.1.0
    # via jsonschema
babel==2.12.1
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
beautifulsoup4==4.12.2
    # via nbconvert
bleach==6.0.0
    # via nbconvert
boto3==1.26.159
    # via -r dummy.requirements.in
botocore==1.29.159
    # via
    #   boto3
    #   s3transfer
build==0.10.0
    # via pip-tools
cerberus==1.3.4
    # via -r dummy.requirements.in
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via pip-tools
comm==0.1.3
    # via ipykernel
cryptography==41.0.2
    # via -r dummy.requirements.in
curlify==2.2.1
    # via -r dummy.requirements.in
debugpy==1.6.7
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
distlib==0.3.6
    # via virtualenv
dnspython==2.3.0
    # via pymongo
et-xmlfile==1.1.0
    # via openpyxl
executing==1.2.0
    # via stack-data
falcon==3.1.1
    # via -r dummy.requirements.in
fastjsonschema==2.17.1
    # via nbformat
filelock==3.12.2
    # via virtualenv
fqdn==1.5.1
    # via jsonschema
gunicorn==20.1.0
    # via -r dummy.requirements.in
identify==2.5.24
    # via pre-commit
idna==3.4
    # via
    #   anyio
    #   jsonschema
    #   requests
ipykernel==6.23.2
    # via jupyterlab
ipython==8.14.0
    # via ipykernel
isoduration==20.11.0
    # via jsonschema
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   swagger-ui-py
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
json5==0.9.14
    # via jupyterlab-server
jsonpointer==2.4
    # via jsonschema
jsonschema[format-nongpl]==4.17.3
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jupyter-client==8.2.0
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclient
jupyter-core==5.3.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.6.3
    # via jupyter-server
jupyter-lsp==2.2.0
    # via jupyterlab
jupyter-server==2.6.0
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook-shim
jupyter-server-terminals==0.4.4
    # via jupyter-server
jupyterlab==4.0.2
    # via -r dummy.requirements.in
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-server==2.23.0
    # via jupyterlab
markupsafe==2.1.3
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==3.0.1
    # via nbconvert
mongoengine==0.27.0
    # via -r dummy.requirements.in
nbclient==0.8.0
    # via nbconvert
nbconvert==7.6.0
    # via jupyter-server
nbformat==5.9.0
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.5.6
    # via ipykernel
nodeenv==1.8.0
    # via pre-commit
notebook-shim==0.2.3
    # via jupyterlab
openpyxl==3.1.2
    # via -r dummy.requirements.in
overrides==7.3.1
    # via jupyter-server
packaging==23.1
    # via
    #   build
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
passlib==1.7.4
    # via -r dummy.requirements.in
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.13.0
    # via -r dummy.requirements.in
platformdirs==3.8.0
    # via
    #   jupyter-core
    #   virtualenv
pre-commit==3.3.3
    # via -r dummy.requirements.in
prometheus-client==0.17.0
    # via jupyter-server
prompt-toolkit==3.0.38
    # via ipython
psutil==5.9.5
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pycryptodome==3.18.0
    # via -r dummy.requirements.in
pygments==2.15.1
    # via
    #   ipython
    #   nbconvert
pyjwt==2.7.0
    # via -r dummy.requirements.in
pymongo==4.4.0
    # via mongoengine
pyproject-hooks==1.0.0
    # via build
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   jupyter-client
python-dotenv==1.0.0
    # via -r dummy.requirements.in
python-json-logger==2.0.7
    # via jupyter-events
pyyaml==6.0
    # via
    #   jupyter-events
    #   pre-commit
    #   swagger-ui-py
pyzmq==25.1.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
requests==2.31.0
    # via
    #   -r dummy.requirements.in
    #   curlify
    #   jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
s3transfer==0.6.1
    # via boto3
send2trash==1.8.2
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.0
    # via anyio
soupsieve==2.4.1
    # via beautifulsoup4
stack-data==0.6.2
    # via ipython
swagger-ui-py==22.7.13
    # via -r dummy.requirements.in
terminado==0.17.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.2.1
    # via nbconvert
tornado==6.3.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   terminado
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
typing-extensions==4.6.3
    # via async-lru
uri-template==1.3.0
    # via jsonschema
urllib3==1.26.16
    # via
    #   botocore
    #   requests
virtualenv==20.23.1
    # via pre-commit
wcwidth==0.2.6
    # via prompt-toolkit
webcolors==1.13
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.6.0
    # via jupyter-server
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
